// types/resources.d.ts
export type ProfileRes = {
  UserId?: boolean,
  EngName?: boolean,
  IsMale?: boolean,
  IsStudent?: boolean,
  MajorYear?: boolean,
  DateOfBirth?: boolean,
  PlaceOfBirth?: boolean,
  IsTaiwanese?: boolean,
  NationalId?: boolean,
  Nationality?: boolean,
  PassportNumber?: boolean,
  Address?: boolean,
  EmergencyContactName?: boolean,
  EmergencyContactPhone?: boolean,
  EmergencyContactMobile?: boolean,
  BeneficiaryName?: boolean,
  BeneficiaryRelation?: boolean,
  RiceAmount?: boolean,
  FoodPreference?: boolean,
}

export type MinProfileRes = {
  UserId?: boolean,
  Name?: boolean,
  MobileNumber?: boolean,
  PhoneNumber?: boolean,
}